Create a custom mat-dice-tooltip css class
to avoid hitting the enrichment check on chat message processing per d&d 5e 3.x+ (see #6). Consequently, hand copy a bunch of styles from 5e and core to get a similar visual effect without being able to take advantage of a bunch of existing style rules. This certainly isn't the best way to address the issue, but gets us back to a working state without waiting for a big refactor.